+++ src/test/kotlin/net/pwall/json/JSONSerializerTest.kt
@@ -50,9 +50,9 @@ import java.time.ZoneOffset
 import java.time.ZonedDateTime
 import java.util.BitSet
 import java.util.Calendar
-import java.util.Enumeration
 import java.util.TimeZone
 import java.util.UUID
+import kotlin.test.assertFailsWith
 
 class JSONSerializerTest {
 
@@ -217,16 +217,6 @@ class JSONSerializerTest {
             addValue("Kotlin")
         }
         expect(expected) { JSONSerializer.serialize(ListEnum(list)) }
-    }
-
-    class ListEnum<T>(private val list: List<T>) : Enumeration<T> {
-
-        private var index = 0
-
-        override fun hasMoreElements(): Boolean = index < list.size
-
-        override fun nextElement(): T = if (hasMoreElements()) list[index++] else throw NoSuchElementException()
-
     }
 
     @Test fun `List of String should return JSONArray`() {
@@ -707,6 +697,17 @@ class JSONSerializerTest {
             putValue("number", 2.0)
         }
         expect(expected) { JSONSerializer.serialize(Const(2.0), config) }
+    }
+
+    @Test fun `should fail on use of circular reference`() {
+        val circular1 = Circular1()
+        val circular2 = Circular2()
+        circular1.ref = circular2
+        circular2.ref = circular1
+        val exception = assertFailsWith<JSONException> {
+            JSONSerializer.serialize(circular1)
+        }
+        expect("Circular reference: field ref in Circular2") { exception.message }
     }
 
     private fun intEquals(a: Int, b: Int): Boolean {