+++ src/test/kotlin/net/pwall/json/JSONStringifyTest.kt
@@ -49,6 +49,7 @@ import java.util.UUID
 
 import net.pwall.json.JSONStringify.appendJSON
 import net.pwall.json.test.JSONExpect.Companion.expectJSON
+import kotlin.test.assertFailsWith
 
 class JSONStringifyTest {
 
@@ -275,7 +276,7 @@ class JSONStringifyTest {
 
     @Test fun `should stringify the results of an enumeration`() {
         val list = listOf("tahi", "rua", "toru", "wh\u0101")
-        expect("""["tahi","rua","toru","wh\u0101"]""") { JSONStringify.stringify(JSONSerializerTest.ListEnum(list)) }
+        expect("""["tahi","rua","toru","wh\u0101"]""") { JSONStringify.stringify(ListEnum(list)) }
     }
 
     @Test fun `should stringify a map of string to string`() {
@@ -342,13 +343,13 @@ class JSONStringifyTest {
         expect("\"$str\"") { JSONStringify.stringify(localDateTime) }
     }
 
-    @Test fun `should stringify a OffsetTime`() {
+    @Test fun `should stringify an OffsetTime`() {
         val str = "10:15:06.543+10:00"
         val offsetTime = OffsetTime.parse(str)
         expect("\"$str\"") { JSONStringify.stringify(offsetTime) }
     }
 
-    @Test fun `should stringify a OffsetDateTime`() {
+    @Test fun `should stringify an OffsetDateTime`() {
         val str = "2020-04-10T10:15:06.543+10:00"
         val offsetDateTime = OffsetDateTime.parse(str)
         expect("\"$str\"") { JSONStringify.stringify(offsetDateTime) }
@@ -600,6 +601,17 @@ class JSONStringifyTest {
             property("field2", null)
             property("field3", "gamma")
         }
+    }
+
+    @Test fun `should fail on use of circular reference`() {
+        val circular1 = Circular1()
+        val circular2 = Circular2()
+        circular1.ref = circular2
+        circular2.ref = circular1
+        val exception = assertFailsWith<JSONException> {
+            JSONStringify.stringify(circular1)
+        }
+        expect("Circular reference: field ref in Circular2") { exception.message }
     }
 
     @Test fun `should append to existing Appendable`() {